+++ @ecomplus/storefront-renderer/functions/index.js
@@ -19,7 +19,7 @@ exports.ssr = (req, res, getCacheControl) => {
 
   const setStatusAndCache = (status, defaultCache) => {
     return res.status(status)
-      .set('X-SSR-ID', `v2/${Math.random()}`)
+      .set('X-SSR-ID', `v2.8/${Math.random()}`)
       .set(
         'Cache-Control',
         (typeof getCacheControl === 'function' && getCacheControl(status)) || defaultCache
@@ -50,7 +50,10 @@ exports.ssr = (req, res, getCacheControl) => {
     if (url.slice(-1) === '/') {
       redirect(url.slice(0, -1))
     } else if (url !== '/404' && (/\/[^/.]+$/.test(url) || /\.x?html$/.test(url))) {
-      redirect(`/404?url=${encodeURIComponent(req.url)}`)
+      setStatusAndCache(404, `public, max-age=${(isLongCache ? 120 : 30)}`)
+        .send('<html><head>' +
+          `<meta http-equiv="refresh" content="0; url=/404?url=${encodeURIComponent(url)}"/>` +
+          '</head><body></body></html>')
     } else {
       setStatusAndCache(404, isLongCache
         ? 'public, max-age=60, s-maxage=86400'
@@ -60,7 +63,6 @@ exports.ssr = (req, res, getCacheControl) => {
   }
 
   return renderer(url)
-
     .then(html => {
       if (html) {
         if (cache) {
@@ -75,7 +77,6 @@ exports.ssr = (req, res, getCacheControl) => {
       }
       return false
     })
-
     .catch(err => {
       fallback()
       console.error(err)