+++ @ecomplus/storefront-template/template/js/netlify-cms/base-config/collections/settings/general.js
@@ -175,10 +175,6 @@ export default ({ baseDir }) => ({
               value: 'spacelab'
             },
             {
-              label: 'Superhero',
-              value: 'superhero'
-            },
-            {
               label: 'United',
               value: 'united'
             },